PS2LINK: final warning fixed, now ps2link module synced with https://github.com/ps2dev/ps2link

+++ modules/debug/ps2link/net_fio.c
@@ -709,7 +709,7 @@ int pko_file_serv(void *argv)
         }
 
         dbgprintf("Client connected from %lx\n",
-                  (int)client_addr.sin_addr.s_addr);
+                  client_addr.sin_addr.s_addr);
 
         remote_pc_addr = client_addr.sin_addr.s_addr;
